refactor: Relocate `statusList` to a common util package
This commit moves the `statusList` (a list of `AppointmentStatus` values) from individual screen files to a new common utility file `StatusListHelper.kt` within the `features/home/src/main/kotlin/eg/edu/cu/csds/icare/home/util/` package.

This change centralizes the `statusList` definition, promoting code reuse and maintainability across different parts of the home feature.

**Affected files:**
- **features/home/src/main/kotlin/eg/edu/cu/csds/icare/home/util/StatusListHelper.kt** (new file)
    - Contains the definition of `statusList`.
- **features/home/src/patient/kotlin/eg/edu/cu/csds/icare/home/screen/home/HomeScreen.kt:**
    - Removed local `statusList` definition.
    - Updated import to use the new common `statusList`.
- **features/home/src/staff/kotlin/eg/edu/cu/csds/icare/home/screen/home/HomeScreen.kt:**
    - Updated import to use the new common `statusList`.
- **features/home/src/staff/kotlin/eg/edu/cu/csds/icare/home/screen/clinic/ClinicianContent.kt:**
    - Updated import to use the new common `statusList`.
    - Corrected import path for `AppointmentCard`.
