+++ features/home/src/main/kotlin/eg/edu/cu/csds/icare/home/util/StatusListHelper.kt
@@ -1,0 +1,11 @@
+package eg.edu.cu.csds.icare.home.util
+
+import eg.edu.cu.csds.icare.core.ui.common.AppointmentStatus
+
+val statusList =
+    listOf(
+        AppointmentStatus.PendingStatus,
+        AppointmentStatus.ConfirmedStatus,
+        AppointmentStatus.CancelledStatus,
+        AppointmentStatus.CompletedStatus,
+    )
